requirements.txt
torch
transformers>=4.36.0
accelerate
sentencepiece
protobuf
google
google-crc32c
grpcio
safetensors
pillow
huggingface-hub
uvicorn
fastapi
pydantic
numpy
